requirements.txt
asgiref==3.8.1
Django==5.1.3
django-environ==0.11.2
django-filter==24.3
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
phonenumbers==8.13.49
psycopg2==2.9.10
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.2
